Remove stegos_config crate
Move Config* structs to corresponding crates.

Closes #463
