+++ Cargo.toml
@@ -12,7 +12,6 @@ path = "src/stegos.rs"
 [workspace]
 members = [
     "blockchain",
-    "config",
     "consensus",
     "crypto",
     "keychain",
@@ -24,7 +23,6 @@ members = [
 
 [dependencies]
 stegos_blockchain = { path = "./blockchain" }
-stegos_config = { path = "./config" }
 stegos_consensus = { path = "./consensus" }
 stegos_crypto = { path = "./crypto" }
 stegos_keychain = { path = "./keychain" }
@@ -47,6 +45,10 @@ rustyline = "2.1"
 lazy_static = "1.1"
 atty = "0.2"
 regex = "1.0"
+serde = "1.0"
+serde_derive = "1.0"
+toml = "0.4"
+rand = "0.6"
 
 [dev-dependencies]
 simple_logger = "1.0"
