+++ blockchain/Cargo.toml
@@ -9,12 +9,10 @@ build = "build.rs"
 [dependencies]
 stegos_keychain = { path = "../keychain" }
 stegos_crypto = { path = "../crypto" }
-stegos_config= { path = "../config" }
 stegos_serialization = {path = "../serialization"}
 rocksdb = "0.11.0"
 byteorder = "1.3.1"
 tokio-timer = "0.2"
-
 chrono = "0.4"
 rand = "0.6"
 log = "0.4"
@@ -22,6 +20,8 @@ failure = "0.1"
 bitvector = "0.1"
 protobuf = "2.2"
 tempdir = "0.3"
+serde = "1.0"
+serde_derive = "1.0"
 
 [dev-dependencies]
 simple_logger = "1.0"
