--- config/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "stegos_config"
-version = "0.1.0"
-authors = ["Stegos AG <info@stegos.cc>"]
-edition = "2018"
-
-[dependencies]
-toml = "0.4"
-serde = "1.0"
-serde_derive = "1.0"
-rand = "0.6"
