+++ keychain/Cargo.toml
@@ -10,8 +10,7 @@ failure = "0.1"
 base64 = "0.9"
 lazy_static = "1.1"
 regex = "1.0"
-
 secp256k1 = { version = "0.12", features = ["rand"] }
-
-stegos_config = { path = "../config" }
 stegos_crypto = { path = "../crypto" }
+serde = "1.0"
+serde_derive = "1.0"
