+++ testing/stegos.toml.j2
@@ -1,5 +1,7 @@
 [general]
 log4rs_config = "stegos-log4rs.toml"
+
+[storage]
 database_path = "testing/node{{NODE_ID}}/database"
 
 [keychain]
