+++ wallet/Cargo.toml
@@ -8,7 +8,6 @@ build = "build.rs"
 [dependencies]
 stegos_crypto = { path = "../crypto" }
 stegos_blockchain = { path = "../blockchain" }
-stegos_config = { path = "../config" }
 stegos_txpool = { path = "../txpool" }
 stegos_network = { path = "../network" }
 stegos_node = { path = "../node" }
